requirements.txt
PyYAML==6.0
pyttsx3==2.90
SpeechRecognition==3.8.1
pyjokes==0.6.0
requests==2.26.0